+++ ChangeLog.md
@@ -3,6 +3,12 @@ XP Compiler ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Merged PR #114: Implements first-class callable syntax: `strlen(...)`
+  now returns a closure which if invoked with a string argument, returns
+  its length. Includes support for static and instance methods as well as
+  indirect references like `$closure(...)` and `self::{$expression}(...)`,
+  see https://wiki.php.net/rfc/first_class_callable_syntax
+
 ## 6.6.0 / 2021-07-10
 
 * Emit null-coalesce operator as `$a ?? $a= expression` instead of as